fix: Input field key for repertoire item
While being on the repertoire page, when changing students, the entry
was always set to the student that was there on load. By adding a key
attribute to the createRepertoireItem component consisting of the holder
type and the hoder ID we guarantee that each time the component is
remounted with the correct holder data.